[5759] Add the "unconfined" option for apparmor in lxc profile
In host system using apparmor, the latter can block some access required
by the localbuilder to work correctly.

More specifically, this is the case for Debian, Ubuntu, Linux Mint and
OpenSUSE.

This is why the installation script is updated in order to add the right
configuration in the lxc profile created when using the script to
install from scratch the container.

Signed-off-by: Armand Bénéteau <armand.beneteau@iot.bzh>